collapse RTK query slices into a single slice (as recommended by docs)
