+++ client/src/features/newManifest/NewManifest.tsx
@@ -8,8 +8,7 @@ import React, { useState } from 'react';
 import { Form } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
-import { useAppSelector } from 'store';
-import { siteByEpaIdSelector } from 'store/profileSlice/profile.slice';
+import { useAppSelector, siteByEpaIdSelector } from 'store';
 
 /**
  * NewManifest component allows a user to create a new electronic manifest.
